requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
flask-wtf
python-dotenv
psycopg2
